pyproject.toml
[project]
name = "Padel"
version = "0.1.0"
description = "Automate padel court booking to solve the problem of being unable to book a court"
authors = [
    {name = "Iñigo Arribillaga", email = "98iarribi@gmail.com"},
]
dependencies = ["black>=25.9.0", "playwright>=1.55.0", "python-dotenv>=1.2.1", "loguru>=0.7.3"]
requires-python = ">=3.12"
readme = "README.md"
license = {text = "MIT"}

[build-system]
requires = ["pdm-backend"]
build-backend = "pdm.backend"


[tool.pdm]
distribution = true


[tool.pdm.scripts]
lint = "black src/ tests/"

[dependency-groups]
test = [
    "pytest>=8.4.2",
    "pytest-bdd>=8.1.0",
]
